Deleted emtpy test classes, better "show" view for Item, Tag controller skeleton for show tags and which items are tagged a certain way

+++ grails-app/domain/Item.groovy
@@ -1,4 +1,4 @@
-class Item {
+class Item {              
 
     Product product
     Address address
@@ -16,12 +16,12 @@ class Item {
 
     /* Business Methods */
 
-    void addRating(int score){
+    void addRating(Integer score){
         totalScore += score
         numberOfVotes += 1
     }
 
-    int checkAverageRating(){
+    Double checkAverageRating(){
         totalScore > 0 ? totalScore/numberOfVotes : 0
     }
